Fix for Travis build, now requirements.txt on master is for python3

+++ bin/travis-build.bash
@@ -22,7 +22,12 @@ else
 fi
 
 python setup.py develop
-pip install -r requirements.txt
+if [ -f requirements-py2.txt ]
+then
+    pip install -r requirements-py2.txt
+else
+    pip install -r requirements.txt
+fi
 pip install -r dev-requirements.txt
 cd -
